Posts - More link: replace angular quotation mark with arrow. Remove duplicated property.
The angular quotation mark is a punctuation character in spanish so it looks odd to use it as a UI element. Despite the fact that the theme is in english originally, it's necessary to update this since this character can't be translated.
This commit replaces it with an arrow, more clear for any language.