Minor: Move site logo closer to the title/menu.
The site logo is located to the left of the site title and menu. Currently, the exact position is determined by the site grid, but it varies depending on the size of the screen. On many larger breakpoints, this leaves the icon sitting a little farther away from the text than I feel is comfortable.

This update moves the site title just a little bit closer to the text.