Update editor margins to match the front end on middle-breakpoints
A followup to #502. Closes #236.

Adjusts editor styles between `600px` and `768px` so that the margins are more closely aligned to the front end.

Note that this doesn't make them 100% exact: the editor has about 10px more space on either side. This is because the editor requires more space for the hover/editor controls. We _could_ match these up exactly by adjusting the front end, but I tend to prefer tighter margins in general and I think this is close enough.