Try: Better position wide and fullwide toolbars.
Toolbars are a bit off when wide and fullwide. This is because the vanilla editor style assumes a centered column, with equal padding left and right when a block goes wide and fullwide.

But TwentyNineteen is smarter than that.

This PR tries to make the experience good, but it needs your thoughts.